+++ functional-tests/functional/dart/test/Listeners_test.dart
@@ -35,15 +35,9 @@ class MessageListener implements StringListener {
 
   @override
   void onStructMessage(StringListenerStringStruct message) {}
-
-  @override
-  release() {}
 }
 
-class RouteImpl implements Route {
-  @override
-  release() {}
-}
+class RouteImpl implements Route {}
 
 class RouteProviderImpl implements RouteProvider {
   static bool setRouteWasRun = false;
@@ -54,9 +48,6 @@ class RouteProviderImpl implements RouteProvider {
     setRouteWasRun = true;
     setRouteCouldCast = route is RouteImpl;
   }
-
-  @override
-  release() {}
 }
 
 void main() {