+++ functional-tests/functional/dart/test/MultiListener_test.dart
@@ -36,9 +36,6 @@ class MultiReceiver implements ReceiverA, ReceiverB {
   void receiveB(String message) {
     log.add("ReceiverB: received from Sender: " + message);
   }
-
-  @override
-  void release() {}
 }
 
 void main() {