+++ gluecodium/src/test/resources/smoke/platform_names/output/dart/lib/src/smoke/wee_listener.dart
@@ -10,9 +10,7 @@ abstract class weeListener {
   ) => weeListener$Lambdas(
     WeeMethodLambda,
   );
-  /// @nodoc
-  @Deprecated("Does nothing")
-  void release() {}
+
   void WeeMethod(String WeeParameter);
 }
 // weeListener "private" section, not exported.
@@ -41,16 +39,14 @@ class weeListener$Lambdas implements weeListener {
   weeListener$Lambdas(
     this.WeeMethodLambda,
   );
-  @override
-  void release() {}
+
   @override
   void WeeMethod(String WeeParameter) =>
     WeeMethodLambda(WeeParameter);
 }
 class weeListener$Impl extends __lib.NativeBase implements weeListener {
   weeListener$Impl(Pointer<Void> handle) : super(handle);
-  @override
-  void release() {}
+
   @override
   void WeeMethod(String WeeParameter) {
     final _WeeMethodFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Void Function(Pointer<Void>, Int32, Pointer<Void>), void Function(Pointer<Void>, int, Pointer<Void>)>('library_smoke_PlatformNamesListener_basicMethod__String'));